prep makefile changes for future image installs
